refactor(ui): split FriendSlopUI.cs (1705 ln) into three partials
FriendSlopUI.cs is now 'partial' and carries only fields, lifecycle
methods (Awake/OnEnable/Update), and the RefreshUi coordinator loop.

Three new partials extract the self-contained concerns:

- FriendSlopUI.BuildUi.cs  (~480 ln): BuildUi, BuildInventoryPanel,
  BuildInventoryPreviewRig, EnsureEventSystem, all widget factories
  (CreatePanel/Text/Button/Input), and shared layout utilities.

- FriendSlopUI.Hud.cs      (~280 ln): LayoutHud, UpdateStaminaBar,
  UpdateHealthBar, UpdateChargeBar, UpdateDeathOverlay,
  UpdateInventoryHud, ShowDamageFlash, ShowLateJoinLoading,
  UpdateLoadingScreen, UpdateSunGlare, UpdateFade.

- FriendSlopUI.Menu.cs     (~310 ln): LayoutMenu, OnRestart/Travel/
  CyclePlanet click handlers, all string formatters (BuildStatusText,
  BuildLobbyQueue, BuildObjectiveHudText, FormatReadableJoinCode, …),
  and UpdateJoinCodePanel / CopyJoinCodeToClipboard.

dotnet build FriendSlop.UI.csproj and FriendSlop.Runtime.csproj both
pass 0 errors after the split.
