+++ Space Game/Assets/Scripts/UI/FriendSlopUI.BuildUi.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: d4e5f6a7b8c94d0eb1f2a3b4c5d6e7f8
