+++ Space Game/Assets/Scripts/UI/FriendSlopUI.Hud.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: e5f6a7b8c9d04e1fb2a3c4d5e6f7a8b9
