+++ Space Game/Assets/Scripts/UI/FriendSlopUI.Menu.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: f6a7b8c9d0e14f2ca3b4c5d6e7f8a9b0
